Division now uses the Double data type

+++ src/main/java/com/qa/day2/Calculator.java
@@ -16,7 +16,7 @@ public class Calculator{
     public static int subtraction(int num1, int num2){
         return num1-num2;
     }
-    public static int division(int num1, int num2){
+    public static double division(double num1, double num2){
         return num1/num2;
     }
     public static int addition(int num1, int num2){
